refactor dto and naming
