+++ blotztask-ui/src/model/ai-assistant-response.ts
@@ -1,0 +1,7 @@
+import { TaskDetailDTO } from "./task-detail-dto";
+
+// Mapped frontend model
+export interface AIAssistantResponse {
+  message: string;
+  tasks: TaskDetailDTO[];
+}
